rename: move ModuleName from __init__ to info
